+++ app/api/auth/agent-connect/login.ts
@@ -2,12 +2,16 @@ import { agentConnectAuthorizeUrl } from '#clients/authentication/agent-connect/
 import { logFatalErrorInSentry } from '#utils/sentry';
 import { setPathFrom } from '#utils/session';
 import withSession from '#utils/session/with-session';
-import { AgentConnectionFailedException } from './callback';
+import { AgentConnectionFailedException } from '../../../../pages/api/auth/agent-connect/callback';
 
 export default withSession(async function loginRoute(req, res) {
   try {
+    // Get the pathFrom from query params or headers
     const pathFrom =
-      (req?.query?.pathFrom as string) || req.headers.referer || '';
+      req.nextUrl.searchParams.get('pathFrom') ||
+      req.headers.get('referer') ||
+      '';
+
     await setPathFrom(req.session, pathFrom);
     const url = await agentConnectAuthorizeUrl(req);
     res.redirect(url);
