+++ app/api/auth/agent-connect/logout-callback.ts
