+++ app/api/auth/agent-connect/logout.ts
