+++ app/api/auth/france-connect/logout-callback.ts
