+++ app/api/auth/france-connect/logout.ts
